split up backend code to reduce length
