+++ Backend/middlewares/auth.js
@@ -1,0 +1,18 @@
+const jwt = require("jsonwebtoken");
+const secretKey = process.env.JWT_SECRET;
+
+const verifyToken = (req, res, next) => {
+  const token = req.cookies.token;
+  if (!token) {
+    return res.status(401).json({ message: "No token, authorization denied" });
+  }
+  try {
+    const decoded = jwt.verify(token, secretKey);
+    req.user = decoded;
+    next();
+  } catch (error) {
+    res.status(401).json({ message: "Token is not valid" });
+  }
+};
+
+module.exports = verifyToken;
